qlib-bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-qlib-bridge"
version = "0.1.0"
description = "Python sidecar bridge between Nova Quant and Microsoft Qlib"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
  "pyqlib>=0.9.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27.0",        # async test client for FastAPI
  "pylint>=3.0.0",
  "ruff>=0.3.0",          # fast linter/formatter
]
models = [
  "torch>=2.0.0",         # only needed for LSTM/Transformer inference
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bridge*", "config*"]
